requirements.txt
pytest
pytest-playwright
pymysql
cryptography
python-dotenv
anyio
pytest-asyncio
pytest-trio
pytest-twisted
twisted
